+++ mesibot-backend/src/controllers/api.ts
@@ -1,11 +1,10 @@
 import { Router } from "express";
-import { playlistRouter } from "./playlist";
-import { songsRouter } from "./songs";
-import { gamesRouter } from "./games";
+import { playlistRouter } from "./party/playlist/playlist";
+import { songsRouter } from "./party/playlist/songs";
+import { partyRouter } from "./party";
 
 export const apiRouter = Router();
 
 apiRouter.use("/playlists", playlistRouter);
 apiRouter.use("/songs", songsRouter);
-
-apiRouter.use("/games", gamesRouter);
+apiRouter.use("/party", partyRouter);
