+++ mesibot-backend/src/controllers/party/games/models/Room.ts
