+++ mesibot-backend/src/controllers/party/index.ts
@@ -1,0 +1,1 @@
+export { partyRouter } from "./party";
