+++ mesibot-backend/src/controllers/party/playlist/index.ts
@@ -1,0 +1,2 @@
+export { playlistRouter } from "./playlist";
+export { songsRouter } from "./songs";
