+++ mesibot-backend/src/services/playlist/clear.ts
@@ -1,0 +1,15 @@
+import { Playlist } from "../../models/Playlist";
+
+export const clear = async (playlistId: string) => {
+  const playlist = await Playlist.findById(playlistId);
+
+  if (!playlist) {
+    return null;
+  }
+
+  playlist.queue = [] as any;
+
+  await playlist.save();
+
+  return playlist;
+};
