+++ mesibot-backend/src/services/playlist/create.ts
@@ -1,0 +1,9 @@
+import { Playlist } from "../../models/Playlist";
+import { StatisticsService } from "../statistics";
+
+export const create = async (title: string) => {
+  const newPlaylist = new Playlist({ title, songs: [], queue: [], played: [], currentPlaying: null });
+  await newPlaylist.save();
+  await StatisticsService.create(newPlaylist._id.toString());
+  return { title: newPlaylist.title, id: newPlaylist._id };
+};
