+++ mesibot-backend/src/services/playlist/getAll.ts
@@ -1,0 +1,7 @@
+import { Playlist } from "../../models/Playlist";
+
+export const getAll = async () => {
+  const playlists = await Playlist.find().select("title _id");
+
+  return playlists;
+};
