+++ mesibot-backend/src/services/playlist/getOne.ts
@@ -1,0 +1,7 @@
+import { Playlist } from "../../models/Playlist";
+
+export const getOne = async (id: string) => {
+  const playlist = await Playlist.findById(id);
+
+  return playlist;
+};
